+++ Plugins/BetterWardrobe.lua
@@ -25,33 +25,7 @@ function betterWardrobe.LoadLocalization(L)
 end
 
 function betterWardrobe.InjectOptions()
-    local optionsTable = {
-        type = "group",
-        name = addon.L["Better Wardrobe"],
-        args = {
-            Loaded = {
-                order = 1, type = "toggle", width = "full",
-                name = addon.L["Loaded"],
-                desc = addon.L["Loaded Desc"],
-                descStyle = "inline",
-                get = function() return IsAddOnLoaded("BetterWardrobe") end,
-                disabled = true
-            },
-            Line = {
-                order = 2, type = "header", width = "full",
-                name = ""
-            },
-            Description = {
-                order = 3, type = "description", width = "full",
-                name = addon.Util.ReplaceVars
-                {
-                    addon.L["Better Wardrobe Desc"],
-                    betterWardrobe = addon.L["Better Wardrobe"]
-                },
-                fontSize = "medium"
-            }
-        }
-    };
-
-    addon.Options.InjectOptionsTable(optionsTable, "BetterWardrobe", "Plugins", "args");
+    KrowiAF_InjectOptions.AddPluginTable("BetterWardrobe", addon.L["Better Wardrobe"], addon.L["Better Wardrobe Desc"], function()
+        return IsAddOnLoaded("BetterWardrobe");
+    end);
 end