+++ Plugins/InstanceAchievementTracker.lua
@@ -12,92 +12,45 @@ function iat.LoadLocalization(L)
     L["Add To Right Click Menu"] = "Add To Right Click Menu";
     L["Add To Right Click Menu Desc"] = "Add an option to the Right Click Menu that will direct you to tactics if {iat} has data for it.";
     L["Instance Achievement Tracker PluginText"] = "|T:1:8|t" .. "- Shortcut to {plugin_iat} if tactics for the achievement are available ({plugin_iat} is {plugin_enabled}).\n";
-
 end
 
 function iat.InjectOptions()
-    local defaults = {};
-    -- defaults.Enabled = true;
-    defaults.AddToRightClickMenu = true;
+    KrowiAF_InjectOptions.AddDefaults("Plugins", "InstanceAchievementTracker", {
+        AddToRightClickMenu = true;
+    });
 
-    addon.Options.InjectDefaults(defaults, "InstanceAchievementTracker", "Plugins");
-
-    local optionsTable = {
-        type = "group",
-        name = addon.L["Instance Achievement Tracker"],
-        args = {
-            Loaded = {
-                order = 1, type = "toggle", width = "full",
-                name = addon.L["Loaded"],
-                desc = addon.L["Loaded Desc"],
-                descStyle = "inline",
-                get = function() return iat.IsLoaded() end,
-                disabled = true
-            },
-            Line = {
-                order = 2, type = "header", width = "full",
-                name = ""
-            },
-            Description = {
-                order = 3, type = "description", width = "full",
-                name = addon.L["Instance Achievement Tracker Desc"]:ReplaceVars
-                {
-                    iat = addon.L["Instance Achievement Tracker"]
-                },
-                fontSize = "medium"
-            },
-            -- Enabled = {
-            --     order = 4, type = "toggle", width = "full",
-            --     name = addon.L["Enable"],
-            --     desc = addon.L["Enable Desc"],
-            --     get = function() return addon.Options.db.Plugins.InstanceAchievementTracker.Enabled; end,
-            --     set = function()
-            --         addon.Options.db.Plugins.InstanceAchievementTracker.Enabled = not addon.Options.db.Plugins.InstanceAchievementTracker.Enabled;
-            --         addon.Options.Debug(addon.L["Instance Achievement Tracker"] .. " " .. addon.L["Enable"], addon.Options.db.Plugins.InstanceAchievementTracker.Enabled);
-            --     end,
-            --     disabled = function() return not iat.IsLoaded(); end
-            -- },
-            AddToRightClickMenu = {
-                order = 5, type = "toggle", width = "full",
-                name = addon.L["Add To Right Click Menu"],
-                desc = addon.Util.ReplaceVars
-                {
-                    addon.L["Add To Right Click Menu Desc"],
-                    iat = addon.L["Instance Achievement Tracker"]
-                },
-                get = function() return addon.Options.db.Plugins.InstanceAchievementTracker.AddToRightClickMenu end,
-                set = function()
-                    addon.Options.db.Plugins.InstanceAchievementTracker.AddToRightClickMenu = not addon.Options.db.Plugins.InstanceAchievementTracker.AddToRightClickMenu;
-                    addon.Options.Debug(addon.L["Instance Achievement Tracker"] .. " " .. addon.L["Enable"], addon.Options.db.Plugins.InstanceAchievementTracker.AddToRightClickMenu);
-                end,
-                disabled = function() return not iat.IsLoaded() --[[or not addon.Options.db.Plugins.InstanceAchievementTracker.Enabled;]] end
-            }
-        }
-    };
-
-    addon.Options.InjectOptionsTable(optionsTable, "InstanceAchievementTracker", "Plugins", "args");
+    local OrderPP = KrowiAF_InjectOptions.AutoOrderPlusPlus;
+    local desc = addon.L["Instance Achievement Tracker Desc"]:ReplaceVars(addon.L["Instance Achievement Tracker"]);
+    local pluginTable = KrowiAF_InjectOptions.AddPluginTable("InstanceAchievementTracker", addon.L["Instance Achievement Tracker"], desc, function()
+        return iat.IsLoaded();
+    end);
+    KrowiAF_InjectOptions.AddTable(pluginTable, "AddToRightClickMenu", {
+        order = OrderPP(), type = "toggle", width = "full",
+        name = addon.L["Add To Right Click Menu"],
+        desc = addon.L["Add To Right Click Menu Desc"]:ReplaceVars(addon.L["Instance Achievement Tracker"]):AddDefaultValueText("Plugins.InstanceAchievementTracker.AddToRightClickMenu"),
+        get = function() return addon.Options.db.Plugins.InstanceAchievementTracker.AddToRightClickMenu; end,
+        set = function() addon.Options.db.Plugins.InstanceAchievementTracker.AddToRightClickMenu = not addon.Options.db.Plugins.InstanceAchievementTracker.AddToRightClickMenu; end,
+        disabled = function() return not iat.IsLoaded(); end
+    });
 end
 
 function iat.Load()
-    addon.L["Right Click Menu Desc"] = addon.Util.ReplaceVars
-    {
-        addon.L["Right Click Menu Desc"],
+    addon.L["Right Click Menu Desc"] = addon.L["Right Click Menu Desc"]:ReplaceVars{
         pluginText = addon.L["Instance Achievement Tracker PluginText"]
     };
-    addon.L["Right Click Menu Desc"] = addon.Util.ReplaceVars
-    {
-        addon.L["Right Click Menu Desc"],
-        plugin_iat = string.format(addon.Colors.Yellow, addon.L["Instance Achievement Tracker"]),
-        plugin_enabled = (iat.IsLoaded() and (string.format(addon.Colors.Green, addon.L["Enabled"]:lower())) or (string.format(addon.Colors.Red, addon.L["Disabled"]:lower())))
+    addon.L["Right Click Menu Desc"] = addon.L["Right Click Menu Desc"]:ReplaceVars{
+        plugin_iat = addon.L["Instance Achievement Tracker"]:SetColorYellow(),
+        plugin_enabled = (iat.IsLoaded() and (addon.L["Enabled"]:lower():SetColorGreen()) or (addon.L["Disabled"]:lower():SetColorRed()))
     };
 end
 
 function iat:AddRightClickMenuItems(rightClickMenu, achievement)
     if self.IsLoaded()
-    -- and addon.Options.db.Plugins.InstanceAchievementTracker.Enabled
     and addon.Options.db.Plugins.InstanceAchievementTracker.AddToRightClickMenu
     and IAT_HasAchievement(achievement.ID) then
-		rightClickMenu:AddFull({Text = addon.L["IAT Tactics"], Func = function() IAT_DisplayAchievement(achievement.ID); end});
+		rightClickMenu:AddFull({Text = addon.L["IAT Tactics"], Func = function()
+            IAT_DisplayAchievement(achievement.ID);
+        end});
 	end
 end